Switch to setuptools and add cythonize to build
Distutils is not fully supported by pip anymore and on newer versions of
pip packages installed that are using distuils can't be uninstalled.
This can be avoided by switching to setuptools to handle the packaging,
which this commit does.

Also, to enable working on newer versions of python this commit
integrates cythonize to the build workflow. This way regardless of
versions the c source file will be generated at build time by cython
and then compiled instead of manually generating them before the build.
The only tradeoff with this is that Cython becomes a build dependency
and must be installed prior to trying to build python-lirc.